Make reshape a member function of ndarray
Fix #824
