+++ pythran/pythonic/numpy/ravel.hpp
@@ -3,7 +3,7 @@
 
 #include "pythonic/include/numpy/ravel.hpp"
 
-#include "pythonic/numpy/reshape.hpp"
+#include "pythonic/numpy/ndarray/reshape.hpp"
 #include "pythonic/utils/numpy_conversion.hpp"
 
 PYTHONIC_NS_BEGIN
@@ -13,7 +13,7 @@ namespace numpy
   template <class T, size_t N>
   types::ndarray<T, 1> ravel(types::ndarray<T, N> const &expr)
   {
-    return reshape(expr, types::array<long, 1>{{expr.flat_size()}});
+    return expr.reshape(types::array<long, 1>{{expr.flat_size()}});
   }
 
   NUMPY_EXPR_TO_NDARRAY0_IMPL(ravel);
